Modify and pin requirements versions for integration with LBUC.
Modify the version requirements in requirements.txt to pin specific
versions of libraries for incubator repository, and remove any
requirements which conflict with packages which conflict with system
packages in the Sagemath/LBUC Docker image. This is necessary to use the
incubator code within a LBUC docker image but a better solution will be
required before merging the repository with upstream.